requirements.txt
# Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets

# Google Cloud
google-genai
google-cloud-speech>=2.26.0

# Utils
python-dotenv==1.0.1
loguru
